re-structure files
