+++ combinetf2/fitter.py
@@ -1,386 +1,11 @@
-import copy
-import tensorflow as tf
-import tensorflow_io as tfio
+
+import hist
 import numpy as np
 import scipy
-import h5py
-import hist
-import narf.ioutils
-import os
-
-
-def maketensor(h5dset):
-    if 'original_shape' in h5dset.attrs:
-        shape = h5dset.attrs['original_shape']
-    else:
-        shape = h5dset.shape
-
-    if h5dset.size == 0:
-        return tf.zeros(shape,h5dset.dtype)
-
-    filename = h5dset.file.filename
-    dsetpath = h5dset.name
-
-    atensor = tfio.IOTensor.from_hdf5(filename)(dsetpath).to_tensor()
-
-    atensor = tf.reshape(atensor, shape)
-    return atensor
-
-
-def makesparsetensor(h5group):
-    indices = maketensor(h5group['indices'])
-    values = maketensor(h5group['values'])
-    dense_shape = h5group.attrs['dense_shape']
-
-    return tf.sparse.SparseTensor(indices, values, dense_shape)
-
-
-def hist_boost(name, axes, values, variances=None, label=None):
-    if not isinstance(axes, (list, tuple, np.ndarray)):
-        axes = [axes]
-    storage_type = hist.storage.Weight() if variances is not None else hist.storage.Double()
-    h = hist.Hist(*axes, storage=storage_type, name=name, label=label)
-    h.values()[...] = memoryview(tf.reshape(values, h.shape))
-    if variances is not None:
-        h.variances()[...] = memoryview(tf.reshape(variances, h.shape))
-    return h
-
-
-def hist_H5PickleProxy(*args, **kwargs):
-    return narf.ioutils.H5PickleProxy(hist_boost(*args, **kwargs))
-
-
-def load_H5PickleProxy(proxy):
-    return proxy.get()
-
-
-class FitInputData:
-    def __init__(self, filename, pseudodata=None, normalize=False):
-        with h5py.File(filename, mode='r') as f:
-
-            #load text arrays from file
-            self.procs = f['hprocs'][...]
-            self.signals = f['hsignals'][...]
-            self.systs = f['hsysts'][...]
-            self.systsnoprofile = f['hsystsnoprofile'][...]
-            self.systsnoconstraint = f['hsystsnoconstraint'][...]
-            self.systgroups = f['hsystgroups'][...]
-            self.systgroupidxs = f['hsystgroupidxs'][...]
-
-            self.noigroups = f['hnoigroups'][...]
-            self.noigroupidxs = f['hnoigroupidxs'][...]
-            if "hpseudodatanames" in f.keys():
-                self.pseudodatanames = f['hpseudodatanames'][...].astype(str)
-            else:
-                self.pseudodatanames = []
-
-            #load arrays from file
-            hconstraintweights = f['hconstraintweights']
-            hdata_obs = f['hdata_obs']
-
-            if 'hdata_cov_inv' in f.keys():
-                hdata_cov_inv = f['hdata_cov_inv']
-                self.data_cov_inv = maketensor(hdata_cov_inv)
-            else:
-                self.data_cov_inv = None
-
-            self.sparse = not 'hnorm' in f
-
-            if self.sparse:
-                print("WARNING: The sparse tensor implementation is experimental and probably slower than with a dense tensor!")
-                hnorm_sparse = f['hnorm_sparse']
-                hlogk_sparse = f['hlogk_sparse']
-            else:
-                hnorm = f['hnorm']
-                hlogk = f['hlogk']
-
-            #infer some metadata from loaded information
-            self.dtype = hdata_obs.dtype
-            self.nbins = hdata_obs.shape[-1]
-            self.nproc = len(self.procs)
-            self.nsyst = len(self.systs)
-            self.nsystnoprofile = len(self.systsnoprofile)
-            self.nsystnoconstraint = len(self.systsnoconstraint)
-            self.nsignals = len(self.signals)
-            self.nsystgroups = len(self.systgroups)
-            self.nnoigroups = len(self.noigroups)
-
-            # reference meta data if available
-            self.metadata = {}
-            if "meta" in f.keys():
-                from narf.ioutils import pickle_load_h5py
-                self.metadata = pickle_load_h5py(f["meta"])
-                self.channel_info = self.metadata["channel_info"]
-            else:
-                self.channel_info = {
-                    "ch0":{"axes": [hist.axis.Integer(0, self.nbins, underflow=False, overflow=False, name="obs")]}
-                }
-
-            # compute indices for channels
-            ibin = 0
-            for channel, info in self.channel_info.items():
-                axes = info["axes"]
-                shape = tuple([len(a) for a in axes])
-                size = np.prod(shape)
-
-                start = ibin
-                stop = start + size
-
-                info["start"] = start
-                info["stop"] = stop
-
-                ibin = stop
-
-            for channel, info in self.channel_info.items():
-                print(channel, info)
-
-            self.axis_procs = hist.axis.StrCategory(self.procs, name="processes")                
-
-            #build tensorflow graph for likelihood calculation
-
-            #start by creating tensors which read in the hdf5 arrays (optimized for memory consumption)
-            self.constraintweights = maketensor(hconstraintweights)
-
-            #load data/pseudodata
-            if pseudodata is not None:
-                if pseudodata in self.pseudodatanames:
-                    pseudodata_idx = np.where(self.pseudodatanames == pseudodata)[0][0]
-                else:
-                    raise Exception("Pseudodata %s not found, available pseudodata sets are %s" % (pseudodata, self.pseudodatanames))
-                print("Run pseudodata fit for index %i: " % (pseudodata_idx))
-                print(self.pseudodatanames[pseudodata_idx])
-                hdata_obs = f['hpseudodata']
-
-                data_obs = maketensor(hdata_obs)
-                self.data_obs = data_obs[:, pseudodata_idx]
-            else:
-                self.data_obs = maketensor(hdata_obs)
-
-            hkstat = f['hkstat']
-            self.kstat = maketensor(hkstat)
-
-            if self.sparse:
-                self.norm_sparse = makesparsetensor(hnorm_sparse)
-                self.logk_sparse = makesparsetensor(hlogk_sparse)
-            else:
-                self.norm = maketensor(hnorm)
-                self.logk = maketensor(hlogk)
-
-            self.normalize = normalize
-            if self.normalize:
-                # normalize predictoin and each systematic to total event yield in data
-                # FIXME this should be done per-channel ideally
-
-                data_sum = tf.reduce_sum(self.data_obs)
-                norm_sum = tf.reduce_sum(self.norm)
-                lognorm_sum = tf.math.log(norm_sum)[None, None, ...]
-
-                logkavg = self.logk[..., 0, :]
-                logkhalfdiff = self.logk[..., 1, :]
-
-                logkdown = logkavg - logkhalfdiff
-                logdown_sum = tf.math.log(tf.reduce_sum(tf.exp(-logkdown) * self.norm[..., None], axis=(0,1)))[None, None, ...]
-                logkdown = logkdown + logdown_sum - lognorm_sum
-
-                logkup = logkavg + logkhalfdiff
-                logup_sum = tf.math.log(tf.reduce_sum(tf.exp(logkup) * self.norm[..., None], axis=(0,1)))[None, None, ...]
-                logkup = logkup - logup_sum + lognorm_sum
-
-                # Compute new logkavg and logkhalfdiff
-                logkavg = 0.5 * (logkup + logkdown)
-                logkhalfdiff = 0.5 * (logkup - logkdown)
-
-                # Stack logkavg and logkhalfdiff to form the new logk_array using tf.stack
-                logk_array = tf.stack([logkavg, logkhalfdiff], axis=-2)
-
-                # Finally, set self.logk to the new computed logk_array
-                self.logk = logk_array
-                self.norm = self.norm * (data_sum / norm_sum)[None, None, ...]
-    
-    def getImpactsAxes(self):
-        impact_names = list(self.signals.astype(str)) + list(self.systs.astype(str))
-        return hist.axis.StrCategory(impact_names, name="impacts")
-
-    def getGlobalImpactsAxes(self):
-        impact_names = list(self.systs.astype(str)[self.nsystnoconstraint:])
-        return hist.axis.StrCategory(impact_names, name="impacts")
-
-    def getImpactsAxesGrouped(self, bin_by_bin_stat=False):
-        impact_names_grouped = list(self.systgroups.astype(str))
-        # impact data stat
-        impact_names_grouped.append("stat")
-        if bin_by_bin_stat:
-            # impact bin-by-bin stat
-            impact_names_grouped.append("binByBinStat")
-        return hist.axis.StrCategory(impact_names_grouped, name="impacts")
-
-class FitDebugData:
-    def __init__(self, indata):
-
-        if indata.sparse:
-            raise NotImplementedError("sparse mode is not supported yet")
-
-        self.indata = indata
-
-        self.axis_procs = self.indata.axis_procs
-        self.axis_systs = hist.axis.StrCategory(indata.systs, name="systs")
-        self.axis_downup = hist.axis.StrCategory(["Down", "Up"], name="DownUp")
-
-        self.data_obs_hists = {}
-        self.nominal_hists = {}
-        self.syst_hists = {}
-        self.syst_active_hists = {}
-
-        ibin = 0
-        for channel, info in self.indata.channel_info.items():
-            axes = info["axes"]
-            shape = [len(a) for a in axes]
-            stop = ibin+np.prod(shape)
-
-            shape_norm = [*shape, self.indata.nproc]
-            shape_logk = [*shape, self.indata.nproc, 2, self.indata.nsyst]
-
-            data_obs_hist = hist.Hist(*axes, name=f"{channel}_data_obs")
-            data_obs_hist.values()[...] = memoryview(tf.reshape(self.indata.data_obs[ibin:stop], shape))
-
-            nominal_hist = hist.Hist(*axes, self.axis_procs, name=f"{channel}_nominal")
-            nominal_hist.values()[...] = memoryview(tf.reshape(self.indata.norm[ibin:stop,:], shape_norm))
-
-            # TODO do these operations on logk in tensorflow instead of numpy to use
-            # multiple cores
-            logk_array = np.asarray(memoryview(tf.reshape(self.indata.logk[ibin:stop,:], shape_logk)))
-            logkavg = logk_array[..., 0, :]
-            logkhalfdiff = logk_array[..., 1, :]
-
-            syst_hist = hist.Hist(*axes, self.axis_procs, self.axis_systs, self.axis_downup, name=f"{channel}_syst")
-
-            logkdown = logkavg - logkhalfdiff
-            syst_hist[{"DownUp" : "Down"}] = np.exp(-logkdown)*nominal_hist.values()[..., None]
-            del logkdown
-
-            logkup = logkavg + logkhalfdiff
-            syst_hist[{"DownUp" : "Up"}] = np.exp(logkup)*nominal_hist.values()[..., None]
-            del logkup
-
-            syst_active_hist = hist.Hist(self.axis_procs, self.axis_systs, name=f"{channel}_syst_active", storage = hist.storage.Int64())
-            syst_active_hist.values()[...] = np.sum(np.logical_or(np.abs(logkavg) > 0., np.abs(logkhalfdiff) > 0.), axis=tuple(range(len(axes))))
-
-            self.data_obs_hists[channel] = data_obs_hist
-            self.nominal_hists[channel] = nominal_hist
-            self.syst_hists[channel] = syst_hist
-            self.syst_active_hists[channel] = syst_active_hist
-
-            ibin = stop
-
-    def nonzeroSysts(self, channels = None, procs = None):
-        if channels is None:
-            channels = self.indata.channel_info.keys()
-
-        if procs is None:
-            procs = list(self.axis_procs)
-
-        nonzero_syst_idxs = set()
-
-        for channel in channels:
-            syst_active_hist = self.syst_active_hists[channel]
-            syst_active_hist = syst_active_hist[{"processes" : procs}]
-            idxs = np.nonzero(syst_active_hist)
-            syst_axis_idxs = syst_active_hist.axes.name.index("systs")
-            syst_idxs = idxs[syst_axis_idxs]
-            nonzero_syst_idxs.update(syst_idxs)
-
-        nonzero_systs = []
-        for isyst, syst in enumerate(self.indata.systs):
-            if isyst in nonzero_syst_idxs:
-                nonzero_systs.append(syst)
-
-        return nonzero_systs
-
-    def channelsForNonzeroSysts(self, procs = None, systs = None):
-        if procs is None:
-            procs = list(self.axis_procs)
-
-        if systs is None:
-            systs = list(self.axis_systs)
-
-        channels_out = []
-
-        for channel in self.indata.channel_info:
-            syst_active_hist = self.syst_active_hists[channel]
-            syst_active_hist = syst_active_hist[{"processes" : procs, "systs" : systs}]
-            if np.count_nonzero(syst_active_hist.values()) > 0:
-                channels_out.append(channel)
-
-        return channels_out
-
-    def procsForNonzeroSysts(self, channels = None, systs = None):
-        if channels is None:
-            channels = self.indata.channel_info.keys()
-
-        if systs is None:
-            systs = list(self.axis_systs)
-
-        proc_idxs_out = set()
-
-        for channel in channels:
-            syst_active_hist = self.syst_active_hists[channel]
-            syst_active_hist = syst_active_hist[{"systs" : systs}]
-            idxs = np.nonzero(syst_active_hist)
-            proc_axis_idxs = syst_active_hist.axes.name.index("processes")
-            proc_idxs = idxs[proc_axis_idxs]
-            proc_idxs_out.update(proc_idxs)
-
-        nonzero_procs = []
-        for iproc, proc in enumerate(self.indata.procs):
-            if iproc in proc_idxs_out:
-                nonzero_procs.append(proc)
-
-        return nonzero_procs
-
-
-class Workspace:
-    def __init__(self, output_format="narf"):
-        self.output_format = output_format
-
-        if output_format == "narf":
-            self.hist = hist_H5PickleProxy
-            self.pack = narf.ioutils.H5PickleProxy
-            self.unpack = load_H5PickleProxy
-            self.dump = narf.ioutils.pickle_dump_h5py
-            self.extension="hdf5"
-        elif output_format == "h5py":
-            self.hist = hist_boost
-            self.pack = lambda x: x
-            self.unpack = lambda x: x
-            self.dump = lambda name, obj, fout: fout.create_dataset(name, obj)
-            self.extension="h5"
-            raise NotImplementedError(f"Writing to vanilla h5py is under development and not yet supported")
-        else:
-            raise ValueError(f"Unknown output format {output_format}")
-
-    def project(self, h, axes):
-        h = self.unpack(h).project(*axes)
-        return self.pack(h)
-
-    def write(self, output, results, meta={}):
-        outfolder = os.path.dirname(output)
-        if outfolder:
-            if not os.path.exists(outfolder):
-                print(f"Creating output folder {outfolder}")
-                os.makedirs(outfolder)
-
-        if '.' not in output and output.split('.')[-1]:
-            output += f".{self.extension}"
-
-        print(f"Write output file {output}")
-
-        with h5py.File(output, "w") as fout:
-            self.dump("results", results, fout)
-            self.dump("meta", meta, fout)
-
+import tensorflow as tf
 
 class Fitter:
-    def __init__(self, indata, options, workspace=None):
+    def __init__(self, indata, workspace, options):
         self.indata = indata
         self.binByBinStat = options.binByBinStat
         self.normalize = options.normalize
@@ -446,7 +71,7 @@ class Fitter:
         nexpfullcentral = self.expected_events(profile=False)
         self.nexpnom = tf.Variable(nexpfullcentral, trainable=False, name="nexpnom")
 
-        self.workspace = Workspace() if workspace is None else workspace
+        self.workspace = workspace
         self.hist = self.workspace.hist
 
     def prefit_covariance(self, unconstrained_err=0.):
